% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readgmt.R
\name{get_gmt_names}
\alias{get_gmt_names}
\title{Get gene set name}
\usage{
get_gmt_names(gl)
}
\arguments{
\item{gl}{a line of the GMT file}
}
\value{
the first value in the list (i.e. the gene set name)
}
\description{
Returns the gene set name from a line of the GMT file
}
