% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readgmt.R
\name{parse_gmt_lines}
\alias{parse_gmt_lines}
\title{Get genes of a gene set}
\usage{
parse_gmt_lines(gl)
}
\arguments{
\item{gl}{a line of the GMT file}
}
\value{
a list of the genes
}
\description{
Returns a list of the genes from a single line in the GMT file
}
