% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readgmt.R
\name{read_gmt}
\alias{read_gmt}
\title{Read a GMT file from MSigDB}
\usage{
read_gmt(file)
}
\arguments{
\item{file}{path to the downloaded GMT file}
}
\value{
returns a named list of the gene set
}
\description{
Reads a GMT file containing a gene set from
    \href{http://software.broadinstitute.org/gsea/index.jsp}{MSigDB} and
    returns a list of named gene sets
}
\examples{
kegg_path <- system.file("extdata", "c2.cp.kegg.v6.2.symbols.gmt.txt",
                         package = "readgmt")
kegg <- read_gmt(kegg_path)
head(names(kegg))
head(kegg$KEGG_COLORECTAL_CANCER)

}
