% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_gmt.R
\name{tidy_gmt}
\alias{tidy_gmt}
\title{Convert the GMT to a "tidy" tibble}
\usage{
tidy_gmt(gmt)
}
\arguments{
\item{gmt}{a gmt from \code{read_gmt()}}
}
\value{
a tibble with columns \code{"gene_set"} and \code{"gene"}
}
\description{
Sometimes, the data is wanted in a "tidy" format with each piece
  of information as an individual row. \code{tidy_gmt()} converts the list
  output of \code{read_gmt()} to a tibble.
}
\examples{
# TODO

}
